src/core/crystal/refraction.py
"""Refraction System Implementation for Crystal Structures.

This module implements the refraction system for crystal analysis:
- Symmetry computation and detection
- Advanced lattice detection algorithms
- Brillouin zone analysis
- Band structure computation
- Phonon mode analysis
"""

import torch
import torch.nn as nn
from typing import Dict, List, Optional, Tuple, Union
from dataclasses import dataclass
import numpy as np
from ..quantum.crystal import BravaisLattice, BrillouinZone, BlochFunction
from ..quantum.state_space import QuantumState, HilbertSpace

@dataclass
class SymmetryOperation:
    """Represents a symmetry operation in the crystal."""
    matrix: torch.Tensor       # Transformation matrix
    translation: torch.Tensor  # Translation vector
    order: int                # Order of the operation
    type: str                 # Type of symmetry (rotation, reflection, etc.)

@dataclass
class BandStructure:
    """Represents the band structure along high-symmetry paths."""
    energies: torch.Tensor    # Energy eigenvalues
    states: List[BlochFunction]  # Corresponding eigenstates
    k_points: torch.Tensor    # k-points along path
    labels: List[str]         # Labels for high-symmetry points

class SymmetryDetector:
    """Detect and analyze crystal symmetries."""
    
    def __init__(
        self,
        lattice: BravaisLattice,
        tolerance: float = 1e-6
    ):
        self.lattice = lattice
        self.tolerance = tolerance
        
        # Initialize symmetry detection networks
        self.point_detector = nn.Sequential(
            nn.Linear(lattice.dim ** 2, lattice.dim * 4),
            nn.ReLU(),
            nn.Linear(lattice.dim * 4, lattice.dim ** 2)
        )
        
        self.translation_detector = nn.Sequential(
            nn.Linear(lattice.dim, lattice.dim * 2),
            nn.ReLU(),
            nn.Linear(lattice.dim * 2, lattice.dim)
        )
    
    def detect_point_symmetries(
        self,
        pattern: torch.Tensor
    ) -> List[SymmetryOperation]:
        """Detect point symmetries in crystal pattern."""
        operations = []
        
        # Analyze pattern structure
        pattern_features = pattern.reshape(-1)
        symmetry_matrices = self.point_detector(pattern_features)
        symmetry_matrices = symmetry_matrices.reshape(
            -1, self.lattice.dim, self.lattice.dim
        )
        
        # Identify valid symmetry operations
        for matrix in symmetry_matrices:
            # Check if operation preserves lattice
            if torch.allclose(
                matrix @ self.lattice.direct_vectors,
                self.lattice.direct_vectors,
                atol=self.tolerance
            ):
                # Determine order and type
                order = self._compute_order(matrix)
                sym_type = self._classify_symmetry(matrix)
                
                operations.append(SymmetryOperation(
                    matrix=matrix,
                    translation=torch.zeros(self.lattice.dim),
                    order=order,
                    type=sym_type
                ))
        
        return operations
    
    def _compute_order(self, matrix: torch.Tensor) -> int:
        """Compute order of symmetry operation."""
        current = torch.eye(self.lattice.dim)
        for i in range(1, 13):  # Check up to order 12
            current = current @ matrix
            if torch.allclose(current, torch.eye(self.lattice.dim)):
                return i
        return -1
    
    def _classify_symmetry(self, matrix: torch.Tensor) -> str:
        """Classify type of symmetry operation."""
        det = torch.det(matrix)
        trace = torch.trace(matrix)
        
        if torch.allclose(matrix, torch.eye(self.lattice.dim)):
            return 'identity'
        elif torch.allclose(det, torch.tensor(1.0)):
            if torch.allclose(trace, torch.tensor(-1.0)):
                return 'rotation'
            else:
                return 'rotoinversion'
        else:
            return 'reflection'

class LatticeDetector:
    """Advanced lattice detection and analysis."""
    
    def __init__(
        self,
        dim: int,
        max_vectors: int = 6
    ):
        self.dim = dim
        self.max_vectors = max_vectors
        
        # Lattice detection network
        self.detector = nn.Sequential(
            nn.Linear(dim * max_vectors, dim * max_vectors * 2),
            nn.ReLU(),
            nn.Linear(dim * max_vectors * 2, dim * dim)
        )
    
    def detect_lattice(
        self,
        points: torch.Tensor
    ) -> BravaisLattice:
        """Detect Bravais lattice from point pattern."""
        # Extract lattice vectors
        features = points[:self.max_vectors].reshape(-1)
        basis_matrix = self.detector(features)
        basis_matrix = basis_matrix.reshape(self.dim, self.dim)
        
        # Determine lattice type
        lattice_type = self._determine_lattice_type(basis_matrix)
        
        return BravaisLattice(
            dim=self.dim,
            lattice_type=lattice_type
        )
    
    def _determine_lattice_type(
        self,
        basis: torch.Tensor
    ) -> str:
        """Determine type of Bravais lattice."""
        angles = self._compute_angles(basis)
        lengths = torch.norm(basis, dim=1)
        
        # Check for cubic
        if torch.allclose(angles, torch.tensor(90.0)) and \
           torch.allclose(lengths, lengths[0]):
            return 'cubic'
        # Check for hexagonal
        elif self.dim == 2 and \
             torch.allclose(angles[0], torch.tensor(120.0)):
            return 'hexagonal'
        else:
            return 'triclinic'
    
    def _compute_angles(
        self,
        basis: torch.Tensor
    ) -> torch.Tensor:
        """Compute angles between lattice vectors."""
        normalized = basis / torch.norm(basis, dim=1, keepdim=True)
        cos_angles = torch.mm(normalized, normalized.t())
        angles = torch.acos(cos_angles.clamp(-1, 1))
        return torch.rad2deg(angles)

class BandStructureAnalyzer:
    """Analyze band structure and phonon modes."""
    
    def __init__(
        self,
        brillouin_zone: BrillouinZone,
        num_bands: int = 4
    ):
        self.brillouin_zone = brillouin_zone
        self.num_bands = num_bands
        
        # Band structure computation
        self.band_computer = nn.Sequential(
            nn.Linear(brillouin_zone.lattice.dim, num_bands * 2),
            nn.ReLU(),
            nn.Linear(num_bands * 2, num_bands)
        )
        
        # Phonon mode analysis
        self.phonon_analyzer = nn.Sequential(
            nn.Linear(num_bands, num_bands * 2),
            nn.Tanh(),
            nn.Linear(num_bands * 2, num_bands)
        )
    
    def compute_band_structure(
        self,
        k_path: torch.Tensor
    ) -> BandStructure:
        """Compute band structure along k-path."""
        # Compute energies
        energies = self.band_computer(k_path)
        
        # Compute corresponding states
        states = []
        for k_point in k_path:
            bloch = BlochFunction(
                self.brillouin_zone.lattice,
                HilbertSpace(self.num_bands)
            )
            states.append(bloch)
        
        return BandStructure(
            energies=energies,
            states=states,
            k_points=k_path,
            labels=self._get_path_labels(k_path)
        )
    
    def analyze_phonons(
        self,
        band_structure: BandStructure
    ) -> torch.Tensor:
        """Analyze phonon modes from band structure."""
        return self.phonon_analyzer(band_structure.energies)
    
    def _get_path_labels(
        self,
        k_path: torch.Tensor
    ) -> List[str]:
        """Get labels for high-symmetry points."""
        labels = []
        for k_point in k_path:
            # Check against known high-symmetry points
            for name, point in self.brillouin_zone.high_symmetry_points.items():
                if torch.allclose(k_point, point):
                    labels.append(name)
                    break
            else:
                labels.append('')
        return labels

class RefractionSystem:
    """Complete refraction system for crystal analysis."""
    
    def __init__(
        self,
        dim: int,
        max_vectors: int = 6,
        num_bands: int = 4
    ):
        self.lattice_detector = LatticeDetector(dim, max_vectors)
        self.symmetry_detector = None  # Initialized after lattice detection
        self.band_analyzer = None      # Initialized after lattice detection
    
    def analyze_pattern(
        self,
        pattern: torch.Tensor
    ) -> Tuple[BravaisLattice, List[SymmetryOperation], BandStructure]:
        """Complete analysis of crystal pattern."""
        # Detect lattice
        lattice = self.lattice_detector.detect_lattice(pattern)
        
        # Initialize remaining components
        self.symmetry_detector = SymmetryDetector(lattice)
        brillouin_zone = BrillouinZone(lattice)
        self.band_analyzer = BandStructureAnalyzer(brillouin_zone)
        
        # Perform analysis
        symmetries = self.symmetry_detector.detect_point_symmetries(pattern)
        k_path = self._generate_k_path(brillouin_zone)
        band_structure = self.band_analyzer.compute_band_structure(k_path)
        
        return lattice, symmetries, band_structure
    
    def _generate_k_path(
        self,
        brillouin_zone: BrillouinZone
    ) -> torch.Tensor:
        """Generate k-path through high-symmetry points."""
        points = list(brillouin_zone.high_symmetry_points.values())
        path = []
        
        for i in range(len(points) - 1):
            segment = torch.linspace(0, 1, 10)[:, None] * (points[i+1] - points[i])
            segment = segment + points[i]
            path.append(segment)
        
        return torch.cat(path)
